requirements.txt
# Core dependencies
pydantic>=2.0.0
python-dotenv>=1.0.0
httpx>=0.24.0
aiohttp>=3.8.0

# LLM SDKs
openai>=1.0.0
anthropic>=0.18.0
google-generativeai>=0.3.0

# Vector DB
turbopuffer>=0.1.0

# Sandbox (E2B)
e2b-code-interpreter>=0.0.5

# MCP
mcp>=0.1.0

# Observability
braintrust>=0.0.100

# Utilities
rich>=13.0.0  # Better terminal output
